Renamed list-view to list (mimics Patternfly’s list-pf), in order to distinguish between the different HTML/CSS layouts
