--- src/app/list-view/list-view-actions.component.less
@@ -1,1 +1,0 @@
-@import (reference) "../../assets/stylesheets/patternfly-ng";
