+++ src/app/list/examples/basic-content.component.ts
