+++ src/app/list/examples/hosts-content.component.ts
