+++ src/app/list/examples/list-basic-example.component.ts
@@ -8,35 +8,31 @@ import {
 import { Action } from '../../models/action';
 import { ActionConfig } from '../../models/action-config';
 import { EmptyStateConfig } from '../../empty-state/empty-state-config';
-import { ListViewConfig } from '../list-view-config';
-import { ListViewEvent } from '../list-view-event';
+import { ListConfig } from '../list-config';
+import { ListEvent } from '../list-event';
 
 @Component({
   encapsulation: ViewEncapsulation.None,
-  selector: 'list-view-heading-example',
-  styleUrls: ['./list-view-heading-example.component.less'],
-  templateUrl: './list-view-heading-example.component.html'
+  selector: 'list-basic-example',
+  styleUrls: ['./list-basic-example.component.less'],
+  templateUrl: './list-basic-example.component.html'
 })
-export class ListViewHeadingExampleComponent implements OnInit {
-  actionConfig: ActionConfig;
+export class ListBasicExampleComponent implements OnInit {
   actionsText: string = '';
   allItems: any[];
+  // dragItem: any;
   emptyStateConfig: EmptyStateConfig;
   items: any[];
-  listViewConfig: ListViewConfig;
+  itemsAvailable: boolean = true;
+  listConfig: ListConfig;
   selectType: string = 'checkbox';
+  showDisabledRows: boolean = false;
 
   constructor() {
   }
 
   ngOnInit(): void {
     this.allItems = [{
-      // First array item used for heading
-      name: 'NAME',
-      actions: 'ACTIONS',
-      additionalInfo: 'ADDITOINAL INFO',
-      address: 'ADDRESS'
-    }, {
       name: 'Fred Flintstone',
       address: '20 Dinosaur Way',
       city: 'Bedrock',
@@ -120,12 +116,70 @@ export class ListViewHeadingExampleComponent implements OnInit {
     }];
     this.items = this.allItems;
 
-    this.actionConfig = {
+    this.emptyStateConfig = {
+      actions: {
+        primaryActions: [{
+          id: 'action1',
+          title: 'Main Action',
+          tooltip: 'Start the server'
+        }],
+        moreActions: [{
+          id: 'action2',
+          title: 'Secondary Action 1',
+          tooltip: 'Do the first thing'
+        }, {
+          id: 'action3',
+          title: 'Secondary Action 2',
+          tooltip: 'Do something else'
+        }, {
+          id: 'action4',
+          title: 'Secondary Action 3',
+          tooltip: 'Do something special'
+        }]
+      } as ActionConfig,
+      icon: 'pficon-warning-triangle-o',
+      title: 'No Items Available',
+      info: 'This is the Empty State component. The goal of a empty state pattern is to provide a good first ' +
+        'impression that helps users to achieve their goals. It should be used when a list is empty because no ' +
+        'objects exists and you want to guide the user to perform specific actions.',
+      helpLink: {
+        hypertext: 'List example',
+        text: 'For more information please see the',
+        url: '#/list'
+      }
+    } as EmptyStateConfig;
+
+    this.listConfig = {
+      dblClick: false,
+      emptyStateConfig: this.emptyStateConfig,
+      multiSelect: false,
+      selectItems: false,
+      selectionMatchProp: 'name',
+      showCheckbox: true,
+      useExpandingRows: false
+    } as ListConfig;
+  }
+
+  ngDoCheck(): void {
+  }
+
+  /**
+   * Get the ActionConfig properties for each row
+   *
+   * @param item The current row item
+   * @param actionButtonTemplate {TemplateRef} Custom button template
+   * @param startButtonTemplate {TemplateRef} Custom button template
+   * @returns {ActionConfig}
+   */
+  getActionConfig(item: any, actionButtonTemplate: TemplateRef<any>,
+      startButtonTemplate: TemplateRef<any>): ActionConfig {
+    let actionConfig = {
       primaryActions: [{
         id: 'start',
         styleClass: 'btn-primary',
         title: 'Start',
-        tooltip: 'Start the server'
+        tooltip: 'Start the server',
+        template: startButtonTemplate
       }, {
         id: 'action1',
         title: 'Action 1',
@@ -137,7 +191,8 @@ export class ListViewHeadingExampleComponent implements OnInit {
       }, {
         id: 'action3',
         title: 'Action 3',
-        tooltip: 'Do something special'
+        tooltip: 'Do something special',
+        template: actionButtonTemplate
       }],
       moreActions: [{
         id: 'moreActions1',
@@ -173,19 +228,25 @@ export class ListViewHeadingExampleComponent implements OnInit {
       moreActionsVisible: true
     } as ActionConfig;
 
-    this.listViewConfig = {
-      dblClick: false,
-      emptyStateConfig: this.emptyStateConfig,
-      multiSelect: false,
-      selectItems: false,
-      selectionMatchProp: 'name',
-      showCheckbox: true,
-      useExpandingRows: false,
-      useHeading: true
-    } as ListViewConfig;
-  }
+    // Set button disabled
+    if (item.started === true) {
+      actionConfig.primaryActions[0].disabled = true;
+    }
 
-  ngDoCheck(): void {
+    // Set custom properties for row
+    if (item.name === 'John Smith') {
+      actionConfig.moreActionsStyleClass = 'red'; // Set kebab option text red
+      actionConfig.primaryActions[1].visible = false; // Hide first button
+      actionConfig.primaryActions[2].disabled = true; // Set last button disabled
+      actionConfig.primaryActions[3].styleClass = 'red'; // Set last button text red
+      actionConfig.moreActions[0].visible = false; // Hide first kebab option
+    }
+
+    // Hide kebab
+    if (item.name === 'Frank Livingston') {
+      actionConfig.moreActionsVisible = false;
+    }
+    return actionConfig;
   }
 
   // Actions
@@ -197,38 +258,74 @@ export class ListViewHeadingExampleComponent implements OnInit {
     this.actionsText = $event.title + ' selected\r\n' + this.actionsText;
   }
 
-  handleSelect($event: ListViewEvent): void {
+  handleSelect($event: ListEvent): void {
     this.actionsText = $event.item.name + ' selected\r\n' + this.actionsText;
   }
 
-  handleSelectionChange($event: ListViewEvent): void {
+  handleSelectionChange($event: ListEvent): void {
     this.actionsText = $event.selectedItems.length + ' items selected\r\n' + this.actionsText;
   }
 
-  handleClick($event: ListViewEvent): void {
+  handleClick($event: ListEvent): void {
     this.actionsText = $event.item.name + ' clicked\r\n' + this.actionsText;
   }
 
-  handleDblClick($event: ListViewEvent): void {
+  handleDblClick($event: ListEvent): void {
     this.actionsText = $event.item.name + ' double clicked\r\n' + this.actionsText;
   }
 
-  handleCheckboxChange($event: ListViewEvent): void {
+  handleCheckboxChange($event: ListEvent): void {
     this.actionsText = $event.item.name + ' checked: ' + $event.item.selected + '\r\n' + this.actionsText;
   }
 
+/* Not implemented
+  // Drag and drop
+
+  handleDragEnd($event: ListEvent): void {
+    this.actionsText = 'drag end\r\n' + this.actionsText;
+  }
+
+  handleDragMoved($event: ListEvent): void {
+    let index = -1;
+
+    for (let i = 0; i < this.items.length; i++) {
+      if (this.items[i] === this.dragItem) {
+        index = i;
+        break;
+      }
+    }
+    if (index >= 0) {
+      this.items.splice(index, 1);
+    }
+    this.actionsText = 'drag moved\r\n' + this.actionsText;
+  }
+
+  handleDragStart($event: ListEvent): void {
+    this.dragItem = $event.item;
+    this.actionsText = $event.item.name + ': drag start\r\n' + this.actionsText;
+  }
+*/
+
   // Row selection
+
+  updateDisabledRows(): void {
+    this.items[1].disabled = this.showDisabledRows;
+  }
+
+  updateItemsAvailable(): void {
+    this.items = (this.itemsAvailable) ? this.allItems : [];
+  }
 
   updateSelectionType(): void {
     if (this.selectType === 'checkbox') {
-      this.listViewConfig.selectItems = false;
-      this.listViewConfig.showCheckbox = true;
+      this.listConfig.selectItems = false;
+      this.listConfig.showCheckbox = true;
     } else if (this.selectType === 'row') {
-      this.listViewConfig.selectItems = true;
-      this.listViewConfig.showCheckbox = false;
+      this.listConfig.selectItems = true;
+      this.listConfig.showCheckbox = false;
     } else {
-      this.listViewConfig.selectItems = false;
-      this.listViewConfig.showCheckbox = false;
+      this.listConfig.selectItems = false;
+      this.listConfig.showCheckbox = false;
     }
   }
 }
