+++ src/app/list/list.component.less
@@ -27,7 +27,7 @@
 
 // For displaying a heading above the list
 .pfng-list-heading {
-  // Hide heading for small views
+  // Hide heading for small screens
   @media (max-width: 992px) {
     display: none;
   }
