+++ src/demo/app-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 
 import { EmptyStateExampleComponent } from '../app/empty-state/examples/empty-state-example.component';
 import { FilterExampleComponent } from '../app/filter/examples/filter-example.component';
-import { ListViewExampleComponent } from '../app/list-view/examples/list-view-example.component';
+import { ListExampleComponent } from '../app/list/examples/list-example.component';
 import { NotificationServiceExampleComponent } from '../app/notification/examples/notification-service-example.component';
 import { RemainingCharsCountExampleComponent } from '../app/remaining-chars-count/examples/remaining-chars-count-example.component';
 import { SampleExampleComponent } from '../app/sample/examples/sample-example.component';
@@ -24,8 +24,8 @@ const routes: Routes = [{
     path: 'filters',
     component: FilterExampleComponent
   }, {
-    path: 'listview',
-    component: ListViewExampleComponent
+    path: 'list',
+    component: ListExampleComponent
   }, {
     path: 'notificationservice',
     component: NotificationServiceExampleComponent
