+++ src/demo/app.component.html
@@ -23,7 +23,7 @@
           <a routerLink="/filters" routerLinkActive="active">Filter</a>
         </li>
         <li>
-          <a routerLink="/listview" routerLinkActive="active">List View</a>
+          <a routerLink="/list" routerLinkActive="active">List</a>
         </li>
         <li>
           <a routerLink="/sample" routerLinkActive="active">Sample</a>
